refactor(policies): improve policy controls UX — intuitive, scannable, minimal
