+++ packages/app-core/src/components/policy-controls/ApprovedAddressesSection.tsx
@@ -1,0 +1,173 @@
+import { Button, Input } from "@miladyai/ui";
+import { Plus, Trash2 } from "lucide-react";
+import { useCallback, useMemo, useState } from "react";
+import { isValidAddress } from "./helpers";
+import type { ApprovedAddressEntry, ApprovedAddressesConfig } from "./types";
+
+export function ApprovedAddressesSection({
+  config,
+  onChange,
+}: {
+  config: ApprovedAddressesConfig;
+  onChange: (config: ApprovedAddressesConfig) => void;
+}) {
+  const [newAddress, setNewAddress] = useState("");
+  const [newLabel, setNewLabel] = useState("");
+  const [addressError, setAddressError] = useState<string | null>(null);
+
+  const entries: ApprovedAddressEntry[] = useMemo(
+    () =>
+      (config.addresses ?? []).map((addr) => {
+        if (typeof addr === "object" && addr !== null && "address" in addr) {
+          const obj = addr as unknown as { address: string; label?: string };
+          return { address: obj.address, label: obj.label ?? "" };
+        }
+        return {
+          address: String(addr),
+          label: config.labels?.[String(addr)] ?? "",
+        };
+      }),
+    [config],
+  );
+
+  const handleAdd = useCallback(() => {
+    const trimmed = newAddress.trim();
+    if (!trimmed) return;
+    if (!isValidAddress(trimmed)) {
+      setAddressError("Invalid address (0x + 40 hex chars)");
+      return;
+    }
+    if (config.addresses.includes(trimmed)) {
+      setAddressError("Already in list");
+      return;
+    }
+    onChange({
+      ...config,
+      addresses: [...config.addresses, trimmed],
+      labels: {
+        ...config.labels,
+        ...(newLabel.trim() ? { [trimmed]: newLabel.trim() } : {}),
+      },
+    });
+    setNewAddress("");
+    setNewLabel("");
+    setAddressError(null);
+  }, [newAddress, newLabel, config, onChange]);
+
+  const handleRemove = useCallback(
+    (addr: string) => {
+      const labels = { ...config.labels };
+      delete labels[addr];
+      onChange({
+        ...config,
+        addresses: config.addresses.filter((a) => a !== addr),
+        labels,
+      });
+    },
+    [config, onChange],
+  );
+
+  return (
+    <div className="space-y-3">
+      {/* Mode toggle */}
+      <div className="flex items-center gap-2">
+        <span
+          className={`text-[11px] px-2 py-0.5 rounded-full font-medium ${
+            config.mode === "whitelist"
+              ? "bg-ok/15 text-ok"
+              : "bg-danger/15 text-danger"
+          }`}
+        >
+          {config.mode === "whitelist" ? "Allowlist" : "Blocklist"}
+        </span>
+        <Button
+          variant="ghost"
+          size="sm"
+          className="text-[11px] h-6 px-2"
+          onClick={() =>
+            onChange({
+              ...config,
+              mode: config.mode === "whitelist" ? "blacklist" : "whitelist",
+            })
+          }
+        >
+          Switch to {config.mode === "whitelist" ? "blocklist" : "allowlist"}
+        </Button>
+      </div>
+
+      {/* Address list */}
+      {entries.length > 0 ? (
+        <div className="space-y-1 max-h-[180px] overflow-y-auto">
+          {entries.map((entry) => (
+            <div
+              key={entry.address}
+              className="flex items-center gap-2 rounded-lg bg-bg/50 px-3 py-1.5 group"
+            >
+              <div className="flex-1 min-w-0">
+                <div className="text-[11px] font-mono text-txt truncate">
+                  {entry.address}
+                </div>
+                {entry.label && (
+                  <div className="text-[10px] text-muted">{entry.label}</div>
+                )}
+              </div>
+              <Button
+                variant="ghost"
+                size="sm"
+                className="h-6 w-6 p-0 opacity-0 group-hover:opacity-100 text-danger hover:text-danger"
+                onClick={() => handleRemove(entry.address)}
+                aria-label={`Remove ${entry.label || entry.address}`}
+              >
+                <Trash2 className="h-3 w-3" />
+              </Button>
+            </div>
+          ))}
+        </div>
+      ) : (
+        <div className="text-[11px] text-muted/60 py-1">
+          {config.mode === "whitelist"
+            ? "No addresses — agent can't send anywhere yet."
+            : "No addresses blocked."}
+        </div>
+      )}
+
+      {/* Add new */}
+      <div className="flex gap-2">
+        <Input
+          type="text"
+          placeholder="0x..."
+          value={newAddress}
+          onChange={(e) => {
+            setNewAddress(e.target.value);
+            setAddressError(null);
+          }}
+          className="flex-1 h-8 text-[12px] font-mono"
+        />
+        <Input
+          type="text"
+          placeholder="Label"
+          value={newLabel}
+          onChange={(e) => setNewLabel(e.target.value)}
+          className="w-24 h-8 text-[12px]"
+        />
+        <Button
+          variant="outline"
+          size="sm"
+          className="h-8 px-2.5"
+          onClick={handleAdd}
+        >
+          <Plus className="h-3 w-3" />
+        </Button>
+      </div>
+      {addressError && (
+        <div className="text-[11px] text-danger">{addressError}</div>
+      )}
+    </div>
+  );
+}
+
+export function addressSummary(config: ApprovedAddressesConfig): string {
+  const count = config.addresses?.length ?? 0;
+  const mode = config.mode === "whitelist" ? "allowed" : "blocked";
+  return count === 0 ? `No addresses ${mode}` : `${count} ${mode}`;
+}
