+++ packages/app-core/src/components/policy-controls/AutoApproveSection.tsx
@@ -1,0 +1,52 @@
+import { Input, Label } from "@miladyai/ui";
+import { ShieldCheck } from "lucide-react";
+import { ethToNumber } from "./helpers";
+import type { AutoApproveConfig } from "./types";
+
+export function AutoApproveSection({
+  config,
+  onChange,
+}: {
+  config: AutoApproveConfig;
+  onChange: (config: AutoApproveConfig) => void;
+}) {
+  return (
+    <div className="space-y-3">
+      <div className="space-y-1">
+        <Label className="text-[11px] text-muted">
+          Auto-approve below this amount
+        </Label>
+        <div className="relative w-40">
+          <Input
+            type="text"
+            inputMode="decimal"
+            value={config.threshold}
+            onChange={(e) => {
+              const v = e.target.value;
+              if (/^\d*\.?\d*$/.test(v)) onChange({ threshold: v });
+            }}
+            className="h-8 text-[13px] pr-12 tabular-nums"
+            placeholder="0.01"
+          />
+          <span className="absolute right-3 top-1/2 -translate-y-1/2 text-[11px] text-muted pointer-events-none">
+            ETH
+          </span>
+        </div>
+      </div>
+      <div className="flex items-center gap-2 rounded-lg bg-accent/5 border border-accent/15 px-3 py-1.5">
+        <ShieldCheck className="h-3.5 w-3.5 text-accent shrink-0" />
+        <div className="text-[11px] text-muted">
+          Under{" "}
+          <span className="font-semibold text-txt">
+            {ethToNumber(config.threshold)} ETH
+          </span>{" "}
+          → auto-approved
+        </div>
+      </div>
+    </div>
+  );
+}
+
+export function autoApproveSummary(config: AutoApproveConfig): string {
+  return `< ${config.threshold} ETH`;
+}
