+++ packages/app-core/src/components/policy-controls/SpendingLimitSection.tsx
@@ -1,0 +1,66 @@
+import { Input, Label } from "@miladyai/ui";
+import type { SpendingLimitConfig } from "./types";
+
+function EthInput({
+  label,
+  value,
+  onChange,
+}: {
+  label: string;
+  value: string;
+  onChange: (value: string) => void;
+}) {
+  return (
+    <div className="space-y-1">
+      <Label className="text-[11px] text-muted">{label}</Label>
+      <div className="relative">
+        <Input
+          type="text"
+          inputMode="decimal"
+          value={value}
+          onChange={(e) => {
+            const v = e.target.value;
+            if (/^\d*\.?\d*$/.test(v)) onChange(v);
+          }}
+          className="h-8 text-[13px] pr-12 tabular-nums"
+          placeholder="0.0"
+        />
+        <span className="absolute right-3 top-1/2 -translate-y-1/2 text-[11px] text-muted pointer-events-none">
+          ETH
+        </span>
+      </div>
+    </div>
+  );
+}
+
+export function SpendingLimitSection({
+  config,
+  onChange,
+}: {
+  config: SpendingLimitConfig;
+  onChange: (config: SpendingLimitConfig) => void;
+}) {
+  return (
+    <div className="grid grid-cols-3 gap-3">
+      <EthInput
+        label="Per Transaction"
+        value={config.maxPerTx}
+        onChange={(v) => onChange({ ...config, maxPerTx: v })}
+      />
+      <EthInput
+        label="Daily Max"
+        value={config.maxPerDay}
+        onChange={(v) => onChange({ ...config, maxPerDay: v })}
+      />
+      <EthInput
+        label="Weekly Max"
+        value={config.maxPerWeek}
+        onChange={(v) => onChange({ ...config, maxPerWeek: v })}
+      />
+    </div>
+  );
+}
+
+export function spendingSummary(config: SpendingLimitConfig): string {
+  return `${config.maxPerTx}/tx · ${config.maxPerDay}/day · ${config.maxPerWeek}/wk ETH`;
+}
